CarDetailTest

+++ Entities/Concrete/Car.cs
@@ -17,7 +17,7 @@ namespace Entities.Concrete
 
         public string ModelYear { get; set; }
 
-        public double DailyPrice { get; set; }
+        public int DailyPrice { get; set; }
 
         public string Description { get; set; }
     }
